[yunpengx]  1. add performance measure             2. change morton code to 64bits             3. fix bugs of incorrect pattern             4. update cuda morton space model code             5. update cuda Morton space model, can compare the performance between seq and cuda version

+++ MortonTree.h
@@ -7,7 +7,7 @@
 
 class MortonTreeObject : public Object {
     public:
-        unsigned int mcode;
+        unsigned long mcode;    // 8bytes
         // store the index of cell that contians the MortonTreeObject
         int parent;
